@@ -1,67 +1,60 @@
-namespace DesafioFundamentos.Models
+namespace AppEstacionamento.Models;
+
+public class Estacionamento
 {
-    public class Estacionamento
+    private decimal precoInicial = 0;
+    private decimal precoHora = 0;
+    private List<string> veiculos = new List<string>();
+
+    public Estacionamento(decimal precoInicial, decimal precoHora)
     {
-        private decimal precoInicial = 0;
-        private decimal precoPorHora = 0;
-        private List<string> veiculos = new List<string>();
+        this.precoInicial = precoInicial;
+        this.precoHora = precoHora;
+    }
 
-        public Estacionamento(decimal precoInicial, decimal precoPorHora)
+    //Entrada para digitar a placa do veículo
+    public void AdicionarVeiculo()
+    {
+        Console.WriteLine("Digite a placa do veículo para estacionar:");
+        string placa = Console.ReadLine();
+        veiculos.Add(placa);
+    }
+
+    //Entrada para remover a placa do veículo
+    public void RemoverVeiculo()
+    {
+        Console.WriteLine("Digite a placa do veículo para remover:");
+        //Entrada para o usuário digitar a placa e armazenar a variável
+        string placa = Console.ReadLine();
+
+        //Verificação de existência do veículo
+        if (veiculos.Any(x => x.ToUpper() == placa.ToUpper()))
         {
-            this.precoInicial = precoInicial;
-            this.precoPorHora = precoPorHora;
+            //Solicitar do usuário a quantidade de horas que permaneceu estacionado
+            Console.WriteLine("Digite a quantidade de horas que o veículo permaneceu estacionado");
+            int horas = int.Parse(Console.ReadLine());
+            decimal valorTotal = precoInicial + (horas * precoHora);
+            //Exibir o preço total a ser pago
+            Console.WriteLine($"O veículo {placa} foi removido e o preço total a pagar é de R$ {valorTotal}");
+            veiculos.Remove(placa);
         }
-
-        public void AdicionarVeiculo()
+        else
         {
-            // TODO: Pedir para o usuário digitar uma placa (ReadLine) e adicionar na lista "veiculos"
-            // *IMPLEMENTE AQUI*
-            Console.WriteLine("Digite a placa do veículo para estacionar:");
+            Console.WriteLine("Veículo não encontrado.");
         }
+    }
 
-        public void RemoverVeiculo()
+    public void ListarVeiculos()
+    {
+        if (veiculos.Any())
         {
-            Console.WriteLine("Digite a placa do veículo para remover:");
-
-            // Pedir para o usuário digitar a placa e armazenar na variável placa
-            // *IMPLEMENTE AQUI*
-            string placa = "";
-
-            // Verifica se o veículo existe
-            if (veiculos.Any(x => x.ToUpper() == placa.ToUpper()))
+            Console.WriteLine("Os veículos estacionados são:");
+            foreach (var veiculo in veiculos)
             {
-                Console.WriteLine("Digite a quantidade de horas que o veículo permaneceu estacionado:");
-
-                // TODO: Pedir para o usuário digitar a quantidade de horas que o veículo permaneceu estacionado,
-                // TODO: Realizar o seguinte cálculo: "precoInicial + precoPorHora * horas" para a variável valorTotal                
-                // *IMPLEMENTE AQUI*
-                int horas = 0;
-                decimal valorTotal = 0; 
-
-                // TODO: Remover a placa digitada da lista de veículos
-                // *IMPLEMENTE AQUI*
-
-                Console.WriteLine($"O veículo {placa} foi removido e o preço total foi de: R$ {valorTotal}");
+                Console.WriteLine(veiculo);
             }
-            else
-            {
-                Console.WriteLine("Desculpe, esse veículo não está estacionado aqui. Confira se digitou a placa corretamente");
-            }
-        }
-
-        public void ListarVeiculos()
-        {
-            // Verifica se há veículos no estacionamento
-            if (veiculos.Any())
-            {
-                Console.WriteLine("Os veículos estacionados são:");
-                // TODO: Realizar um laço de repetição, exibindo os veículos estacionados
-                // *IMPLEMENTE AQUI*
-            }
-            else
-            {
-                Console.WriteLine("Não há veículos estacionados.");
-            }
+        }else{
+            Console.WriteLine("Não há veículos estacionados.");
         }
     }
 }
