@@ -1,25 +1,26 @@
-﻿using DesafioFundamentos.Models;
+﻿using AppEstacionamento.Models;
 
-// Coloca o encoding para UTF8 para exibir acentuação
+//Encodign para exibir acentuação
 Console.OutputEncoding = System.Text.Encoding.UTF8;
 
 decimal precoInicial = 0;
-decimal precoPorHora = 0;
+decimal precoHora = 0;
 
-Console.WriteLine("Seja bem vindo ao sistema de estacionamento!\n" +
-                  "Digite o preço inicial:");
+//Entradas Iniciais do Programa
+Console.WriteLine("Sejam bem-vindos(as) ao Park Estacionamento!\n" + "Digite o preço inicial do estacionamento:");
+//Variáveis Iniciais
 precoInicial = Convert.ToDecimal(Console.ReadLine());
 
-Console.WriteLine("Agora digite o preço por hora:");
-precoPorHora = Convert.ToDecimal(Console.ReadLine());
+Console.WriteLine("Agora digite o preço por hora do estacionamento:");
+precoHora = Convert.ToDecimal(Console.ReadLine());
 
-// Instancia a classe Estacionamento, já com os valores obtidos anteriormente
-Estacionamento es = new Estacionamento(precoInicial, precoPorHora);
+//Instância de Classe Estacionamento
+Estacionamento estac = new Estacionamento(precoInicial, precoHora);
 
 string opcao = string.Empty;
 bool exibirMenu = true;
 
-// Realiza o loop do menu
+//Realizar Looping
 while (exibirMenu)
 {
     Console.Clear();
@@ -27,33 +28,26 @@ while (exibirMenu)
     Console.WriteLine("1 - Cadastrar veículo");
     Console.WriteLine("2 - Remover veículo");
     Console.WriteLine("3 - Listar veículos");
-    Console.WriteLine("4 - Encerrar");
+    Console.WriteLine("4 - Encerrar Menu");
 
     switch (Console.ReadLine())
     {
         case "1":
-            es.AdicionarVeiculo();
+            estac.AdicionarVeiculo();
             break;
-
         case "2":
-            es.RemoverVeiculo();
+            estac.RemoverVeiculo();
             break;
-
         case "3":
-            es.ListarVeiculos();
+            estac.ListarVeiculos();
             break;
-
         case "4":
             exibirMenu = false;
             break;
-
         default:
-            Console.WriteLine("Opção inválida");
+            Console.WriteLine("Opção inválida, digite uma opção válida!");
             break;
     }
-
-    Console.WriteLine("Pressione uma tecla para continuar");
-    Console.ReadLine();
+    Console.WriteLine("Pressione qualquer tecla para continuar...");
+    Console.ReadKey();
 }
-
-Console.WriteLine("O programa se encerrou");
